WIP: move zbusdebug
